Added extensions to TwitterJSON class
Trying and group types of methods together to avoid an overly long main
class file.
